Internal class refactoring.
