+++ src/main/Anvil/Internal/ICoreLoopHandler.cs
@@ -1,4 +1,4 @@
-namespace NWN
+namespace Anvil.Internal
 {
   internal interface ICoreLoopHandler
   {
