+++ src/main/NWN/API/Object/NwFaction.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Anvil.Internal;
 using NWN.Native.API;
 using ClassType = NWN.API.Constants.ClassType;
