+++ src/main/NWN/API/Object/NwModule.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Anvil.Internal;
 using NWN.API.Constants;
 using NWN.API.Events;
 using NWN.Core;
