+++ src/main/NWN/NManager.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Anvil.Internal;
 using NLog;
 using NWN.API;
 using NWN.Core;
