+++ src/main/NWN/Plugins/Plugin.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Reflection;
+using Anvil.Internal;
 
 namespace NWN.Plugins
 {
