+++ src/main/NWN/Plugins/PluginLoadContext.cs
@@ -2,6 +2,7 @@ using System;
 using System.IO;
 using System.Reflection;
 using System.Runtime.Loader;
+using Anvil.Internal;
 
 namespace NWN.Plugins
 {
@@ -21,7 +22,7 @@ namespace NWN.Plugins
 
     protected override Assembly Load(AssemblyName assemblyName)
     {
-      if (!NWN.Assemblies.ReservedNames.Contains(assemblyName.Name))
+      if (!Anvil.Internal.Assemblies.ReservedNames.Contains(assemblyName.Name))
       {
         // Try resolving locally from the plugin folder.
         string assemblyPath = resolver.ResolveAssemblyToPath(assemblyName);
