+++ src/main/NWN/Services/GameLoop/LoopService.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using Anvil.Internal;
 using NLog;
 
 namespace NWN.Services
