+++ src/main/NWN/Services/ScriptDispatch/DispatchServiceManager.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Anvil.Internal;
 
 namespace NWN.Services
 {
